refactor-rename packages, move acrawriter into separate package
