+++ wrappers/android_java/AcraWriterAndroidApp/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.cossacklabs.acra">
+    package="com.cossacklabs.acrawriter_example">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivityAcraStructExample">
+        <activity android:name="com.cossacklabs.acrawriter_example.MainActivityAcraStructExample">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
