--- wrappers/android_java/AcraWriterAndroidApp/app/src/main/java/com/cossacklabs/acra/CallbackListener.java
@@ -1,5 +1,0 @@
-package com.cossacklabs.acra;
-
-public interface CallbackListener<T> {
-    void onComplete(T value);
-}
